** BUG FOUND ** > ** BUG FIXED **
It is in semanitic-ui-css when you try to build a new project, there is an extra semicolon in the code causing the project to crash.  Reverting react-scripts to 4.0.3 fixes the issue.  Did this with no breaks in the project as of yet.

Backed out last update to loadData.js which used the controllers.  This caused a Webpack error

Worked out the Header issue with the Client.js file > it has them embedded in the options by default